Powder/auth_service/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose
sqlmodel
pydantic[email]
aiosqlite
sqlalchemy
python-dotenv
python-multipart
